+++ packages/core/src/runtime/core/types/structures.ts
@@ -24,6 +24,7 @@ export interface RillFieldDef {
   name?: string | undefined;
   type: TypeStructure;
   defaultValue?: RillValue | undefined;
+  annotations?: Record<string, RillValue>;
 }
 
 /**